backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = ">=4.2,<5.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
django-guardian = "^2.4.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
ipython = "^9.0.2"
isort = "^6.0.1"
django-debug-toolbar = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
skip = [
"migrations",
"**/migrations",
]